# Phase2/tfm/Swagger/openapi-tspa.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "XFSC-TRAIN",
    "description": "This is the REST API of the XFSC TRAIN TSPA.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:16003",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/tspa/v1/{framework-name}/did": {
      "put": {
        "tags": [
          "trust-frame-work-publish-controller"
        ],
        "operationId": "publishDidUri",
        "parameters": [
          {
            "name": "framework-name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/trust-framework-did"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "trust-frame-work-publish-controller"
        ],
        "operationId": "deleteTrustListDID",
        "parameters": [
          {
            "name": "framework-name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/tspa/v1/trustframework/{framework-name}": {
      "put": {
        "tags": [
          "trust-frame-work-publish-controller"
        ],
        "operationId": "publishTrustFramework",
        "parameters": [
          {
            "name": "framework-name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/trust-framework"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "trust-frame-work-publish-controller"
        ],
        "operationId": "deleteTrustFramework",
        "parameters": [
          {
            "in": "path",
            "name": "framework-name",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Name of trust framework to be deleted"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful removal"
          },
          "400": {
            "description": "Invalid parameters supplied or Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "response": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "trust-framework": {
        "type": "object",
        "properties": {
          "trustframeworkname": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "trust-framework-did": {
        "type": "object",
        "properties": {
          "did": {
            "type": "string"
          }
        }
      },
...
    }
  }
}